fix(desktop): split ManagedAgentProcess out of types.rs
The file-size ratchet blocked the review-fix push: types.rs grew
999 -> 1001 after adding connect_relay_url.
